move op-node http package into op-service httputil + update related imports
