+++ op-node/metrics/metrics.go
@@ -10,7 +10,7 @@ import (
 	"strconv"
 	"time"
 
-	ophttp "github.com/ethereum-optimism/optimism/op-node/http"
+	ophttp "github.com/ethereum-optimism/optimism/op-service/httputil"
 	"github.com/ethereum-optimism/optimism/op-node/p2p/store"
 	"github.com/ethereum-optimism/optimism/op-service/metrics"
